Simple predicate reordering

+++ src/lib/optimizer/statistics.hpp
@@ -10,7 +10,7 @@ namespace opossum {
 
 class Statistics {
  public:
-  static size_t predicate_result_size(const std::shared_ptr<Table> table, const std::string &column_name,
+  static double predicate_result_size(const std::shared_ptr<Table> table, const std::string &column_name,
                                       const std::string &op, const AllParameterVariant value,
                                       const optional<AllTypeVariant> value2 = nullopt);
  private:
